// src/models/member-req.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 会员创建/更新请求
 * @export
 * @interface MemberReq
 */
export interface MemberReq {
    /**
     * 卡类型ID
     * @type {number}
     * @memberof MemberReq
     */
    cardTypeId?: number;
    /**
     * 会员卡号
     * @type {string}
     * @memberof MemberReq
     */
    cardNo?: string;
    /**
     * 姓名
     * @type {string}
     * @memberof MemberReq
     */
    name: string;
    /**
     * 手机号
     * @type {string}
     * @memberof MemberReq
     */
    phone: string;
    /**
     * 身份证号
     * @type {string}
     * @memberof MemberReq
     */
    idCard?: string;
    /**
     * 状态
     * @type {number}
     * @memberof MemberReq
     */
    status?: number;
}