// src/models/recharge-apply-create-req.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 充值申请创建请求
 * @export
 * @interface RechargeApplyCreateReq
 */
export interface RechargeApplyCreateReq {
    /**
     * 会员ID
     * @type {number}
     * @memberof RechargeApplyCreateReq
     */
    memberId: number;
    /**
     * 门店ID
     * @type {number}
     * @memberof RechargeApplyCreateReq
     */
    storeId: number;
    /**
     * 申请人(业务员)ID
     * @type {number}
     * @memberof RechargeApplyCreateReq
     */
    staffId: number;
    /**
     * 充值金额(本金)
     * @type {number}
     * @memberof RechargeApplyCreateReq
     */
    amount: number;
    /**
     * 赠送金额
     * @type {number}
     * @memberof RechargeApplyCreateReq
     */
    giftAmount?: number;
    /**
     * 充值凭证URL列表
     * @type {Array<string>}
     * @memberof RechargeApplyCreateReq
     */
    voucherUrls?: Array<string>;
    /**
     * 备注
     * @type {string}
     * @memberof RechargeApplyCreateReq
     */
    remark?: string;
}